Outsource fetch and refactor code
